CHG: Moved/Renamed files replaced normal datatypes through stdinttypes Moved all files for hardware to driver folder: Header files in /driver. Source files in /driver/src.
Renamed some files with a better name like Verstearker to tpa2016d2.
It should be no possible to find these better and the name descripe for
what this file is.

Most of all standart types like int, char and so on are replaced through
types from the stdint lib like int16_t, int8_t and so on.

Further a little programm is commit which generates a header file based on
the msp430_pin_mapping.h, better description will follow soon.

In the log_exp_table.h are stored a exp. values for brigthness controll.